refactoring SelectionsFragment.kt
*add view model - ready
